win: add support for unicode files
On linux paths are bytestrings and UTF-8 is supported transparently. On
Windows, Unicode paths are a superset of UTF-16. This means that
Smokeview can have issues opening files and directories with non-ASCII
characters in their path.

This adds a conversion function from UTF-8 to UTF-16 and includes it in
the FOPEN, UNLINK, CHDIR, and MKDIR functions.

This also adds tests to verify that this works correctly.

make chdir unicode aware