win: add unicode support to commandline
Previously on Windows when running Smokeview from the commandline, any
path that included Unicode characters could not be opened.

This adds a function for getting the commandline arguments. On
non-Windows platforms this just returns the values unchanged. On Windows
it uses Windows-specific function to retrieve Unicode versions of the
arguments and convert them to an appropriate version.