win: add support for unicode files
On linux paths are bytestrings and UTF-8 is supported transparently. On
Windows, Unicode paths are a superset of UTF-16. This means that
Smokeview can have issues opening files and directories with non-ASCII
characters in their path.

This adds a conversion function from UTF-8 to UTF-16 and includes it in
the FOPEN, UNLINK, CHDIR, and MKDIR functions. The functions are also
moved to have more consistent definitions.

Tests are added to verify that this works correctly.

This also adds Unicode support to commandline arguments.

Previously on Windows when running Smokeview from the commandline, any
path that included Unicode characters could not be opened.

This adds a function for getting the commandline arguments. On
non-Windows platforms this just returns the values unchanged. On Windows
it uses Windows-specific function to retrieve Unicode versions of the
arguments and convert them to an appropriate version.