Gifs (#10)
* add option to create gifs

This is similar to the move making option but produces an anuimated GIF.
It uses the linked GD library rather than calling a separate program.

This can be accessed via the render dialog box (adjacent the movie
options).

There are still improvements to be made namely:

- Adding more options (framerate, render specific frames).
- Adding access to scripts.

GIFs have a limited colour palette, as such stepped colourbars (or
similar) should be used to reduce the number of output colours.

This colour limitation is an issue for detailed results but this is
useful for debugging purposes or simple concepts.

* enable/disable gifs based on libgd version

* place GIF dialog buttons behind libgd version gate

* factor out pp_GIF_ANIMATED and add docs

---------

Co-authored-by: Jake O'Shannessy <joshanessy@threadway.com>